New feature - can add prefix to filename on download from camera
Have added a method of adding a short prefix e.g. '"Client John Beach shoot_ " to the begining of each filename for all downloaded images.   This automatic image naming for a download batch is very useful and saves much labour.   The GUI is getting a bit crowded but  the text box for inputing tis prefix is just above the triangle arrow for starting the image download frm the camera.   Hope tis is helpful to everyone .